Implement XTPUSHSGR, XTPOPSGR
This change adds a new pair of methods to ITermDispatch:
PushGraphicsRendition and PopGraphicsRendition, and then plumbs the
change through AdaptDispatch, TerminalDispatch, ITerminalApi and
TerminalApi.

The stack logic is encapsulated in the SgrStack class, to allow it to be
reused between the two APIs (AdaptDispatch and TerminalDispatch).

Like xterm, only ten levels of nesting are supported.

Pushes beyond ten will remain balanced (an equal number of pops will
take you back down to zero), up to 100 pushes. Beyond 100 pushes, pushes
will become unbalanced (the internal counter will no longer be
incremented). This bound gives the terminal a deterministic way to
recover from garbage--do 101 pops and you know you've cleared the stack
back down to zero.

For partial pushes (see the description of XTPUSHSGR in Issue #1796),
only attributes that are supported by terminal are saved; others are
ignored (this change does not including adding general support for
double underlines, for example). A partial push of unsupported
parameters results in an "empty" push--the subsequent pop will not
change the current text attributes.